提示信息完善

+++ src/main/java/cn/decision01/danmakuvote/Commands/VoteStatusExecutor.java
@@ -20,7 +20,7 @@ public class VoteStatusExecutor implements CommandExecutor {
             return false;
         }
 
-        int nextTimestamp = plugin.nextTimestamp;
+        long nextTimestamp = plugin.nextTimestamp;
         long space = nextTimestamp - System.currentTimeMillis() / 1000;
         String msg;
         if (nextTimestamp == -1) {
